requirements.txt
### dependency
tqdm
dvc
dvc[s3]
pandas
numpy
PyYAML
scikit-learn
Scipy
gdown

### local packages -
-e . # << uncomment this to install from setup.py
